refactor: 抽出 PersonalMemoryPanel 供个人中心与 Embed 共用
